chore: move test music fixtures out of audio module
